DT-633: blt:dev:link-composer lando support (#3773)
* DT-633: dev-link-composer support for Lando.

* code style cleanup.

* Fixed path validation for Lando.

+++ src/Robo/Commands/Blt/DevCommand.php
@@ -11,9 +11,13 @@ use Acquia\Blt\Robo\Common\YamlWriter;
 class DevCommand extends BltTasks {
 
   // Map host BLT path to guest BLT path so that symlinks work in guest.
-  const DIRECTORY_MAPPING = [
+  const VM_DIRECTORY_MAPPING = [
     '../../packages/blt' => '/var/packages/blt',
     '../blt' => '/var/www/blt',
+  ];
+  const LANDO_DIRECTORY_MAPPING = [
+    '../../packages/blt' => '/packages/blt',
+    '../blt' => '/blt',
   ];
 
   /**
@@ -33,39 +37,71 @@ class DevCommand extends BltTasks {
       return;
     }
 
-    // Switch to local BLT version.
-    $this->taskExec("composer config repositories.blt path {$options['blt-path']} && composer require acquia/blt:* --no-update")
-      ->dir($this->getConfigValue('repo.root'))
-      ->run();
+    $symlink = $this->getConfigValue('repo.root') . '/vendor/acquia/blt';
+    if (filetype($symlink) !== 'link') {
+      // Switch to local BLT version.
+      $this->taskExec("composer config repositories.blt path {$options['blt-path']} && composer require acquia/blt:* --no-update")
+        ->dir($this->getConfigValue('repo.root'))
+        ->run();
 
-    // Remove any patches.
-    $this->taskExec("composer config --unset extra.patches.acquia/blt")
-      ->dir($this->getConfigValue('repo.root'))
-      ->run();
+      // Remove any patches.
+      $this->taskExec("composer config --unset extra.patches.acquia/blt")
+        ->dir($this->getConfigValue('repo.root'))
+        ->run();
 
-    // Nuke and reinitialize Composer to pick up changes.
-    $this->taskExec('rm -rf vendor && composer update acquia/blt --with-dependencies')
-      ->dir($this->getConfigValue('repo.root'))
-      ->run();
+      // Nuke and reinitialize Composer to pick up changes.
+      $this->taskExec('rm -rf vendor && composer update acquia/blt --with-dependencies')
+        ->dir($this->getConfigValue('repo.root'))
+        ->run();
+    }
 
     // Mount local BLT in DrupalVM.
-    $projectDrupalVmConfigFile = $this->getConfigValue('vm.config');
-    if ($projectDrupalVmConfigFile && isset(self::DIRECTORY_MAPPING[$options['blt-path']])) {
-      $yamlWriter = new YamlWriter($projectDrupalVmConfigFile);
+    if ($this->getInspector()->isDrupalVmConfigPresent()) {
+      if (!isset(self::VM_DIRECTORY_MAPPING[$options['blt-path']])) {
+        $this->logger->info('BLT path is not valid for usage with DrupalVM.');
+        return;
+      }
+      $yamlWriter = new YamlWriter($this->getConfigValue('vm.config'));
       $vm_config = $yamlWriter->getContents();
       $existing_entry = array_filter($vm_config['vagrant_synced_folders'], function ($folder) {
-        return in_array($folder['destination'], self::DIRECTORY_MAPPING);
+        return in_array($folder['destination'], self::VM_DIRECTORY_MAPPING);
       });
       if ($existing_entry) {
         return;
       }
       $vm_config['vagrant_synced_folders'][] = [
         'local_path' => $options['blt-path'],
-        'destination' => self::DIRECTORY_MAPPING[$options['blt-path']],
+        'destination' => self::VM_DIRECTORY_MAPPING[$options['blt-path']],
         'type' => 'nfs',
       ];
       $yamlWriter->write($vm_config);
       $this->taskExec('vagrant halt && vagrant up')
+        ->dir($this->getConfigValue('repo.root'))
+        ->run();
+    }
+
+    // Mount local BLT in Lando.
+    if ($this->getInspector()->isLandoConfigPresent()) {
+      if (!isset(self::LANDO_DIRECTORY_MAPPING[$options['blt-path']])) {
+        $this->logger->info('BLT path is not valid for usage with Lando.');
+        return;
+      }
+      $yamlWriter = new YamlWriter($this->getConfigValue('repo.root') . '/.lando.yml');
+      $lando_config = $yamlWriter->getContents();
+      if (isset($lando_config['services']['appserver']['overrides']['volumes'])) {
+        $existing_entry = array_filter($lando_config['services']['appserver']['overrides']['volumes'], function ($folder) {
+          list(, $dest) = explode(':', $folder);
+          return in_array($dest, self::LANDO_DIRECTORY_MAPPING);
+        });
+        if ($existing_entry) {
+          return;
+        }
+      }
+      $lando_config['services']['appserver']['overrides']['volumes'] = [
+        $options['blt-path'] . ':' . self::LANDO_DIRECTORY_MAPPING[$options['blt-path']],
+      ];
+      $yamlWriter->write($lando_config);
+      $this->taskExec('lando rebuild -y')
         ->dir($this->getConfigValue('repo.root'))
         ->run();
     }
